snowman-python-lib/requirements.txt
pylint==2.15.7
autopep8==2.0.0
snowflake-connector-python==2.8.3
pandas==1.5.2

# below here needs synced with .devcontainer/docker-scripts/snowflake-connector-install/snowflake-connector_v2.8.3_tested_requirements_310.reqs
asn1crypto==1.5.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.4
filelock==3.8.0
idna==3.4
oscrypto==1.3.0
pycparser==2.21
pycryptodomex==3.16.0
PyJWT==2.6.0
pyOpenSSL==22.1.0
pytz==2022.6
requests==2.28.1
typing_extensions==4.4.0
urllib3==1.26.13
snowflake-connector-python==2.8.3
